move staticctapi in its own package
